Remove the ability to specify integer modifiers to log functions
Only the modifier type can be specified as template arguments to the log functions.
This was done to make the code more maintainable by now only having one variant of each log function.
This change is non-breaking and backwards compatible with code that previously used the enum values in the color or style namespace as modifiers. Those enum values are now types, but have the same name and are thus compatible.